feat: decompose profile page to components
