+++ src/components/profile-form/index.js
@@ -1,0 +1,3 @@
+import ProfileForm from './profile-form';
+
+export default ProfileForm;
