+++ src/components/profile-form/profile-form.module.css
@@ -1,0 +1,22 @@
+.formControls {
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  padding: 0;
+  margin: 0;
+  border: none;
+  visibility: hidden;
+}
+
+.cancelButton {
+  background-color: transparent;
+  color: #4C4CFF;
+  border: none;
+  box-sizing: border-box;
+  padding: 16px 8px;
+  margin-right: 20px;
+}
+
+.cancelButton:hover {
+  cursor: pointer;
+}
