+++ src/components/profile-sidebar/index.js
@@ -1,0 +1,3 @@
+import ProfileSidebar from './profile-sidebar';
+
+export default ProfileSidebar;
