+++ src/components/profile-sidebar/profile-sidebar.module.css
@@ -1,0 +1,23 @@
+.navigation {
+  max-width: 320px;
+}
+
+.linkList {
+  list-style: none;
+  padding: 0;
+  margin: 0 0 80px;
+}
+
+.listItem {
+  height: 64px;
+  display: flex;
+  align-items: center;
+}
+
+.link {
+  text-decoration: none;
+}
+
+.active {
+  color: #F2F2F3;
+}
