+++ src/pages/index.js
@@ -1,9 +1,19 @@
-import HomePage from './home';
-import LoginPage from './login';
-import RegistrationPage from './registration';
-import ForgotPasswordPage from './forgot-password';
-import ResetPasswordPage from './reset-password';
-import ProfilePage from './profile';
-import NotFoundPage from './not-found';
+import HomePage from "./home";
+import LoginPage from "./login";
+import RegistrationPage from "./registration";
+import ForgotPasswordPage from "./forgot-password";
+import ResetPasswordPage from "./reset-password";
+import ProfilePage from "./profile";
+import OrdersPage from "./orders";
+import NotFoundPage from "./not-found";
 
-export { HomePage, LoginPage, RegistrationPage, ForgotPasswordPage, ResetPasswordPage, ProfilePage, NotFoundPage };
+export {
+  HomePage,
+  LoginPage,
+  RegistrationPage,
+  ForgotPasswordPage,
+  ResetPasswordPage,
+  ProfilePage,
+  OrdersPage,
+  NotFoundPage,
+};
