+++ src/pages/orders/index.js
@@ -1,0 +1,3 @@
+import OrdersPage from './orders';
+
+export default OrdersPage;
